Fix cdash warnings

+++ modules/core/test/math/testColVector.cpp
@@ -485,7 +485,7 @@ int main()
     std::cout << "** Test conversion to/from std::vector" << std::endl;
     std::vector<double> std_vector(5);
     for (size_t i = 0; i < std_vector.size(); i++) {
-      std_vector[i] = i;
+      std_vector[i] = (double) i;
     }
     vpColVector v(std_vector);
     if (test("v", v, std_vector) == false)
